Initialized classes to display items in RecyclerView
Created an instance Datasource and called the
loadAffirmations method on it. Stored the result as
myDataset.
Used findViewById() to find a reference to the
RecyclerView within the layout and store it.
Created and ItemAdapter instance, passed two param and
assigned the ItemAdapter object to the adapter property of the recycler view.